MAPSO/mapso-microservice/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
PyPDF2==3.0.1
msoffcrypto-tool==4.12.0
olefile==0.46
python-docx==0.8.11
loguru==0.7.0
pytest==7.4.0
httpx==0.24.1